misc: add a xcb patch to help detect freeze conditions
Patch xcb to expose the internal buffer of a xcb_connection_t, so picom can
assert that it's empty before it goes to sleep. With nix its easy to
setup a dev environment that builds libxcb with this patch.

Also rename the `useClangProfile` devShell to `useClangDebuggable` which
I think is a better name. (Pretty sure I am the only person that uses
this anyways).

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>